Fix calculating the minute length (again).
It turns out we need two methods:
- one to calculate the length of _this_ (past) minute
- one to calculate the length of the _next_ minute

decode_time() and increase_second() both want the length of the next minute.

Also in decode_time(), calculate the length of the next minute _before_ calling add_minute() as the latter changes the minute value (obviously) which invalidates the calculation.

Test both values in the unit tests.