Bugfix: count the number of seconds correctly, i.e. 0..(minute_length-1)
For live decoding, we need to know the previous second value, as increase_second() must be called there *before* decode_time() and handle_edge(), to prevent writing the bits one slot early (i.e. bit 20 would go to the slot of bit 19).

Introduce a new enum DecodeType for this purpose which is used by the constructor, drop the Default constructor.

Clear first_minute in decode_time() instead of increase_second() as it makes more sense to do so there.

Update unit tests for the above changes.